requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.9
numpy==1.18.1
pandas==1.0.1
pd==0.0.1
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
selenium==3.141.0
Shapely==1.7.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
